Create bindinputs file and move methods around
